feat(admin): add pagination to movie list and change in from integer to long

+++ src/main/java/ua/lviv/bas/cinema/dao/CinemaHallRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import ua.lviv.bas.cinema.domain.CinemaHall;
 
-public interface CinemaHallRepository extends JpaRepository<CinemaHall, Integer> {
+public interface CinemaHallRepository extends JpaRepository<CinemaHall, Long> {
 
 }
