+++ src/main/java/ua/lviv/bas/cinema/domain/User.java
@@ -22,7 +22,7 @@ public class User {
 
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Integer id;
+	private Long id;
 
 	@Column(nullable = false, unique = true)
 	private String email;
@@ -84,7 +84,7 @@ public class User {
 		this.userRole = userRole;
 	}
 
-	public User(Integer id, String email, String firstName, String lastName, LocalDate dateOfBirth, String city,
+	public User(Long id, String email, String firstName, String lastName, LocalDate dateOfBirth, String city,
 			String phoneNumber, String password, String passwordConfirm, UserRole userRole) {
 		this.id = id;
 		this.email = email;
@@ -98,11 +98,11 @@ public class User {
 		this.userRole = userRole;
 	}
 
-	public Integer getId() {
+	public Long getId() {
 		return id;
 	}
 
-	public void setId(Integer id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
